+++ src/main/java/racingcar/application/validation/Validator.java
@@ -1,4 +1,4 @@
-package racingcar.io.validation;
+package racingcar.application.validation;
 
 public interface Validator {
