+++ src/test/java/racingcar/application/validation/CountValidatorTest.java
@@ -1,11 +1,10 @@
-package racingcar.io.validation;
+package racingcar.application.validation;
 
 import static org.junit.jupiter.api.Assertions.assertDoesNotThrow;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
-import racingcar.io.validation.CountValidator;
 
 class CountValidatorTest {
